refactor(agent): collapse initializeAgent into AgentRuntime.addAgent
initializeAgent was a free function with a single caller (createRuntime)
that always followed it with setAgentTools + registerAgent. The split
between init and runtime was a historical artifact (CLI/TUI shared usage
that no longer exists). Folding the steps into a runtime method makes
the API one-stop and eliminates a redundant module.

- Add AgentRuntime.addAgent(opts) — does workspace prep, tool creation,
  setAgentTools, registerAgent in one call.
- createRuntime loops with runtime.addAgent() instead of the previous
  initializeAgent → setAgentTools → registerAgent triplet.
- Delete src/agent/init.ts.
